requirements.txt
python-telegram-bot==20.1
requests==2.28.1
TikTokApi==1.0.7
